make `Spec` public, make `Format` public
